==> 5_SumStats_Processing.R <==
#Script takes the full results generated by the python agent-based model "main_mm5.py" 
#and creates summary statistics by date and number of open desks. 

library(lubridate)
library(ggplot2)
library(dplyr)
library(tidyverse)
library(xts)
library(plyr)

#Unhash the required date for processing. 
#'scn_bad' indicates the scenarios that need to be removed due to unfeasible landing schedules

start_date = ""
end_date = ""
date_text = ""
scn_bad = c()

setwd("")

#loop through open desk number
for(i in 6:15) {
  i = toString(10)
  statement = paste("Running ",i," desks analysis.")
  print(statement)
  start = as.POSIXct(start_date)
  end = as.POSIXct(end_date)
  periods = as.POSIXct(seq.POSIXt(start,end, by=900))     #creates sequence of 15 minute periods for required date
  per_frame = data.frame(Period = periods)
  path = paste(date_text, "/full_results_", date_text,"_",i,"_500.csv", sep="") 
  print("Loading File")
  ps = read.csv(path,quote="", na.strings = "", stringsAsFactors = FALSE)
  print("File Loaded")
  #filter out bad scenarios
  ps = filter(ps, !Scenario %in% scn_bad)
  ps$Exit_Time = as.POSIXct(ps$Exit_Time)
  
  #calculate the 15 minute period for the exit time
  print("Calculating Period")
  ps$Period <- align.time(ps$Exit_Time- 899, n = 60*15)
  
  #Generate various statistic
  print("Generating Stats")
  #Calculate scenarios with highest and lowest total wait time
  s_tot <- aggregate(WaitTime_Mins ~ Scenario, ps, sum)
  s_tot <- rename(s_tot, c("WaitTime_Mins" = "total_wait_mins"))
  s_max = s_tot %>% top_n(1, total_wait_mins)
  s_max_no =  as.integer(s_max$Scenario)
  s_min = s_tot %>% top_n(-1, total_wait_mins)
  s_min_no =  as.integer(s_min$Scenario)
  
  #Calcuate the average wait time for each period in each scenario
  avg_wait_mm <- aggregate(WaitTime_Mins ~ Scenario + Period, ps, mean)
  #isolate 'worst case' using analysis above
  max_avgWait <- filter(avg_wait_mm, Scenario == s_max_no)
  max_avgWait <- select(max_avgWait, Period, Worst_Case_Avg = WaitTime_Mins)
  #isolate 'best case' using analysis above
  min_avgWait <- filter(avg_wait_mm, Scenario == s_min_no)
  min_avgWait <- select(min_avgWait, Period, Best_Case_Avg = WaitTime_Mins)
  
  #filter for the scheduled scenario
  scheduled = filter(avg_wait_mm, Scenario== 1)
  scheduled <- select(scheduled, Period, Scheduled = WaitTime_Mins)
  
  #Group ps dataframe by interval period and calculate average wait time in minutes
  avg_wait <- aggregate(WaitTime_Mins ~ Period, ps, mean)
  avg_wait <- avg_wait %>% rename(c("WaitTime_Mins" = "Overall_Avg"))
  #Repeat for maximum wait time in each period across all scenarios
  max_wait = aggregate(WaitTime_Mins ~ Period, ps, max)
  max_wait <- max_wait %>% rename(c("WaitTime_Mins" = "Max_Wait"))
  
  #calculate the maximum wait time in each period FOR each scenario
  max_s_wait = aggregate(WaitTime_Mins ~ Period + Scenario, ps, max)
  #calcuate the number of periods that have a wait time above 45/90 for each scenario
  s_45_tot = max_s_wait %>% filter(WaitTime_Mins > 45) %>% count("Period") %>% rename(c("freq" = "Over45"))
  s_90_tot = max_s_wait %>% filter(WaitTime_Mins > 90) %>% count("Period") %>% rename(c("freq" = "Over90"))
  
  #merge together all of the above statistics with the df of all 15 minute periods
  periods = merge(per_frame, avg_wait, all.x = TRUE)
  periods = merge(periods, max_wait, all.x = TRUE)
  periods = merge(periods, max_avgWait, all.x = TRUE)
  periods = merge(periods, min_avgWait, all.x = TRUE)
  periods = merge(periods, scheduled, all.x = TRUE)
  periods = merge(periods, s_45_tot, all.x = TRUE)
  periods = merge(periods, s_90_tot, all.x = TRUE)
  #Calcuate total number of scenarios that the stats use
  total_scenario = length(unique(ps$Scenario))
  #Fill na values with 0. 
  periods = periods %>% replace(is.na(.), 0)
  #Create columns showing the ratio of scenario periods over 45/90 mins compared to all scenario periods
  periods$Over45 = periods$Over45/total_scenario
  periods$Over90 = periods$Over90/total_scenario
  
  #calcuate 95th percentile wait times for each scenario
  w95 = ddply(ps, "Scenario", summarise, PC95 = quantile(WaitTime_Mins, .95))
  #calculate overal probability that a scenario will have a 95th percentile wait time above 45 minutes
  prob_95 = sum(w95$PC95 >45)/total_scenario
  #calculate average wait time for all scenarios
  avg_wait = mean(ps$WaitTime_Mins)
  #create new columns for above statistics
  periods$prob_breachSLA = prob_95
  periods$avg_wait = avg_wait
  periods$desk_no = as.numeric(i)
  #save file
  file_name = paste(date_text,"/", date_text, "_" ,i, "_500.csv",sep="")
  print("Saving File")
  write.csv(periods, file_name )
}
